requirements.txt
tensorflow-gpu
scipy
matplotlib
h5py
SeisCL
